requirements.txt
scikit-learn == 1.2.2
matplotlib == 3.9.1
pandas == 2.2.2
numpy == 1.26.4
pandasai
pyyaml